requirements.txt
Flask==3.0.0
tensorflow==2.15.0
Pillow==8.3.2
numpy==1.23.5

